// client-side/src/types/employee.ts
export type Employee = {
  // employeeId: ReactNode;
  // employeeRole: ReactNode;
  // firstName: ReactNode;
  // imagePath: any;
  // _id: any;
  // DP: string; // Display picture (image path)
  // employeeID: string;
  // employeeName: string;
  // contactNo: string;
  // role: string;
  // salary: number;
  _id:string;
  employeeId: string;
  firstName: string;
  lastName: string;
  email:string;
  contactNo: string;
  employeeRole: string;
  employeeSalary: string;
  address:string;
  imagePath: string;
  
};